refactor: implement speculative feature handling
