+++ src/llama.cpp
@@ -17,6 +17,7 @@
 #include "llama-cparams.h"
 #include "llama-hparams.h"
 #include "llama-context.h"
+#include "llama-spec-features.h"
 #include "llama-quantize.h"
 
 #include "unicode.h"
@@ -25,7 +26,6 @@
 #include "ggml-alloc.h"
 #include "ggml-backend.h"
 
-uint32_t llama_mtp_state_n_embd(const struct llama_context * ctx);
 void llama_set_mtp_target_context(struct llama_context * ctx, struct llama_context * target_ctx);
 
 // TODO: fix these includes
@@ -4239,11 +4239,7 @@ static void llama_set_inputs(llama_context & lctx, const llama_batch & batch) {
 // Make sure enough space is available for outputs.
 // Returns max number of outputs for which space was reserved.
 static uint32_t llama_output_embd_width(const llama_context & lctx) {
-    const auto & hparams = lctx.model.hparams;
-    if (lctx.cparams.mtp && lctx.model.arch == LLM_ARCH_GEMMA4_MTP && hparams.mtp_backbone_n_embd > 0) {
-        return hparams.mtp_backbone_n_embd;
-    }
-    return hparams.n_embd;
+    return llama_mtp_state_n_embd(&lctx);
 }
 
 static bool llama_context_has_mtp_outputs(const llama_context & lctx) {
@@ -8900,169 +8896,6 @@ float * llama_get_embeddings(struct llama_context * ctx) {
     return ctx->embd;
 }
 
-static bool llama_spec_prepare_hidden_feature_view(
-        struct llama_context   * ctx,
-        int32_t                  n_rows,
-        llama_spec_feature_view & view) {
-    view.kind = LLAMA_SPEC_FEATURE_HIDDEN_STATE;
-    view.width = 0;
-    view.rows.clear();
-
-    if (ctx == nullptr || n_rows < 0) {
-        return false;
-    }
-
-    llama_synchronize(ctx);
-
-    if (ctx->embd == nullptr) {
-        return false;
-    }
-
-    view.width = (int32_t) llama_output_embd_width(*ctx);
-    if (view.width <= 0 || ctx->n_outputs_embd < n_rows) {
-        view.width = 0;
-        return false;
-    }
-
-    view.rows.reserve(n_rows);
-    return true;
-}
-
-bool llama_spec_get_hidden_feature_view(
-        struct llama_context   * ctx,
-        const llama_batch      & batch,
-        llama_spec_feature_view & view) {
-    if (batch.n_tokens <= 0 || batch.pos == nullptr || batch.n_seq_id == nullptr || batch.seq_id == nullptr) {
-        return false;
-    }
-
-    if (!llama_spec_prepare_hidden_feature_view(ctx, batch.n_tokens, view)) {
-        return false;
-    }
-
-    for (int32_t i = 0; i < batch.n_tokens; ++i) {
-        if (batch.n_seq_id[i] <= 0 || batch.seq_id[i] == nullptr) {
-            view.rows.clear();
-            return false;
-        }
-
-        view.rows.push_back({
-            /* .seq_id = */ batch.seq_id[i][0],
-            /* .pos    = */ batch.pos[i],
-            /* .data   = */ ctx->embd + (size_t) i * view.width,
-        });
-    }
-
-    return true;
-}
-
-bool llama_spec_get_hidden_feature_view_for_seq(
-        struct llama_context   * ctx,
-        const llama_batch      & batch,
-        llama_seq_id             seq_id,
-        llama_spec_feature_view & view) {
-    if (batch.n_tokens <= 0 || batch.pos == nullptr || batch.n_seq_id == nullptr || batch.seq_id == nullptr) {
-        return false;
-    }
-
-    if (!llama_spec_prepare_hidden_feature_view(ctx, batch.n_tokens, view)) {
-        return false;
-    }
-
-    for (int32_t i = 0; i < batch.n_tokens; ++i) {
-        if (batch.n_seq_id[i] <= 0 || batch.seq_id[i] == nullptr) {
-            view.rows.clear();
-            return false;
-        }
-
-        for (int32_t j = 0; j < batch.n_seq_id[i]; ++j) {
-            if (batch.seq_id[i][j] != seq_id) {
-                continue;
-            }
-
-            view.rows.push_back({
-                /* .seq_id = */ seq_id,
-                /* .pos    = */ batch.pos[i],
-                /* .data   = */ ctx->embd + (size_t) i * view.width,
-            });
-            break;
-        }
-    }
-
-    return !view.rows.empty();
-}
-
-bool llama_spec_get_hidden_feature_view_from_output_index(
-        struct llama_context   * ctx,
-        int32_t                  output_index,
-        llama_seq_id             seq_id,
-        llama_pos                pos,
-        llama_spec_feature_view & view) {
-    if (!llama_spec_prepare_hidden_feature_view(ctx, 1, view)) {
-        return false;
-    }
-
-    if (output_index < 0) {
-        output_index += ctx->n_outputs_embd;
-    }
-    if (output_index < 0 || output_index >= ctx->n_outputs_embd) {
-        view.rows.clear();
-        return false;
-    }
-
-    view.rows.push_back({
-        /* .seq_id = */ seq_id,
-        /* .pos    = */ pos,
-        /* .data   = */ ctx->embd + (size_t) output_index * view.width,
-    });
-    return true;
-}
-
-bool llama_spec_copy_hidden_rows_from_output_indices(
-        struct llama_context * ctx,
-        const std::vector<int32_t> & output_indices,
-        std::vector<float> & hidden_rows) {
-    hidden_rows.clear();
-    if (output_indices.empty()) {
-        return false;
-    }
-
-    llama_spec_feature_view view;
-    if (!llama_spec_prepare_hidden_feature_view(ctx, (int32_t) output_indices.size(), view)) {
-        return false;
-    }
-
-    hidden_rows.reserve((size_t) output_indices.size() * view.width);
-    for (int32_t output_index : output_indices) {
-        if (output_index < 0) {
-            output_index += ctx->n_outputs_embd;
-        }
-        if (output_index < 0 || output_index >= ctx->n_outputs_embd) {
-            hidden_rows.clear();
-            return false;
-        }
-
-        const float * row = ctx->embd + (size_t) output_index * view.width;
-        hidden_rows.insert(hidden_rows.end(), row, row + view.width);
-    }
-
-    return hidden_rows.size() == (size_t) output_indices.size() * view.width;
-}
-
-bool llama_set_draft_input_hidden_state_copy(
-        struct llama_context * ctx,
-        const float * hidden_state,
-        size_t n_floats) {
-    if (ctx == nullptr || hidden_state == nullptr || n_floats == 0) {
-        return false;
-    }
-
-    ctx->draft_input_hidden_state_owned.assign(hidden_state, hidden_state + n_floats);
-    ctx->draft_input_hidden_state = ctx->draft_input_hidden_state_owned.data();
-    ctx->draft_input_hidden_state_n_floats = n_floats;
-    return true;
-}
-
 float * llama_get_embeddings_ith(struct llama_context * ctx, int32_t i) {
     int32_t j = -1;
 
@@ -10325,10 +10158,6 @@ void llama_set_draft_input_hidden_state(struct llama_context * ctx, const float 
     ctx->draft_input_hidden_state_n_floats = ctx->inp_mtp_states
         ? ggml_nbytes(ctx->inp_mtp_states) / sizeof(float)
         : 0;
-}
-
-uint32_t llama_mtp_state_n_embd(const struct llama_context * ctx) {
-    return llama_output_embd_width(*ctx);
 }
 
 void llama_set_mtp_target_context(struct llama_context * ctx, struct llama_context * target_ctx) {
